# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Messenger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  online: "online",
  offline: "offline",
  away: "away",
} as const;

export interface User {
  id: number;
  username: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  avatarColor: string;
  status: UserStatus;
  lastSeen?: string | null;
  phoneNumber?: string | null;
  isBot: boolean;
  createdAt: string;
}

export type UpdateUserBodyStatus =
  (typeof UpdateUserBodyStatus)[keyof typeof UpdateUserBodyStatus];

export const UpdateUserBodyStatus = {
  online: "online",
  offline: "offline",
  away: "away",
} as const;

export interface UpdateUserBody {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  status?: UpdateUserBodyStatus;
}

export interface AddContactBody {
  userId: number;
}

export type ChatType = (typeof ChatType)[keyof typeof ChatType];

export const ChatType = {
  direct: "direct",
  group: "group",
  channel: "channel",
} as const;

export type MessageType = (typeof MessageType)[keyof typeof MessageType];

export const MessageType = {
  text: "text",
  image: "image",
  video: "video",
  audio: "audio",
  file: "file",
  gift: "gift",
  call: "call",
  sticker: "sticker",
  system: "system",
} as const;

export interface Reaction {
  id: number;
  messageId: number;
  userId: number;
  emoji: string;
  user?: User | null;
  createdAt: string;
}

export type GiftItemAnimationType =
  (typeof GiftItemAnimationType)[keyof typeof GiftItemAnimationType];

export const GiftItemAnimationType = {
  confetti: "confetti",
  sparkle: "sparkle",
  hearts: "hearts",
  fireworks: "fireworks",
  stars: "stars",
  balloons: "balloons",
  roses: "roses",
  diamonds: "diamonds",
} as const;

export type GiftItemRarity =
  (typeof GiftItemRarity)[keyof typeof GiftItemRarity];

export const GiftItemRarity = {
  common: "common",
  rare: "rare",
  epic: "epic",
  legendary: "legendary",
} as const;

export interface GiftItem {
  id: number;
  name: string;
  emoji: string;
  animationType: GiftItemAnimationType;
  rarity: GiftItemRarity;
  stars: number;
  description: string;
}

export interface Gift {
  id: number;
  giftItemId: number;
  senderId: number;
  receiverId: number;
  message?: string | null;
  isAnonymous: boolean;
  giftItem?: GiftItem | null;
  sender?: User | null;
  receiver?: User | null;
  createdAt: string;
}

export interface Message {
  id: number;
  chatId: number;
  senderId: number;
  text?: string | null;
  type: MessageType;
  mediaUrl?: string | null;
  replyToId?: number | null;
  isEdited: boolean;
  isRead: boolean;
  reactions: Reaction[];
  sender?: User | null;
  replyTo?: Message | null;
  giftData?: Gift | null;
  createdAt: string;
  updatedAt: string;
}

export type ChatMemberRole =
  (typeof ChatMemberRole)[keyof typeof ChatMemberRole];

export const ChatMemberRole = {
  owner: "owner",
  admin: "admin",
  member: "member",
} as const;

export interface ChatMember {
  id: number;
  chatId: number;
  userId: number;
  role: ChatMemberRole;
  joinedAt: string;
  user?: User;
}

export interface Chat {
  id: number;
  type: ChatType;
  name?: string | null;
  description?: string | null;
  avatarUrl?: string | null;
  avatarColor: string;
  isPinned: boolean;
  isMuted: boolean;
  unreadCount: number;
  lastMessage?: Message | null;
  members?: ChatMember[];
  otherUser?: User | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateChatBodyType =
  (typeof CreateChatBodyType)[keyof typeof CreateChatBodyType];

export const CreateChatBodyType = {
  direct: "direct",
  group: "group",
  channel: "channel",
} as const;

export interface CreateChatBody {
  type: CreateChatBodyType;
  name?: string;
  description?: string;
  memberIds?: number[];
}

export interface UpdateChatBody {
  name?: string;
  description?: string;
  avatarUrl?: string;
  isMuted?: boolean;
}

export type AddMemberBodyRole =
  (typeof AddMemberBodyRole)[keyof typeof AddMemberBodyRole];

export const AddMemberBodyRole = {
  admin: "admin",
  member: "member",
} as const;

export interface AddMemberBody {
  userId: number;
  role?: AddMemberBodyRole;
}

export type SendMessageBodyType =
  (typeof SendMessageBodyType)[keyof typeof SendMessageBodyType];

export const SendMessageBodyType = {
  text: "text",
  image: "image",
  video: "video",
  audio: "audio",
  file: "file",
  sticker: "sticker",
} as const;

export interface SendMessageBody {
  chatId: number;
  text?: string;
  type?: SendMessageBodyType;
  mediaUrl?: string;
  replyToId?: number;
}

export interface EditMessageBody {
  text: string;
}

export interface AddReactionBody {
  emoji: string;
}

export type CallType = (typeof CallType)[keyof typeof CallType];

export const CallType = {
  audio: "audio",
  video: "video",
  group: "group",
} as const;

export type CallStatus = (typeof CallStatus)[keyof typeof CallStatus];

export const CallStatus = {
  ringing: "ringing",
  active: "active",
  ended: "ended",
  missed: "missed",
  declined: "declined",
} as const;

export interface Call {
  id: number;
  chatId?: number | null;
  callerId: number;
  calleeId?: number | null;
  type: CallType;
  status: CallStatus;
  startedAt?: string | null;
  endedAt?: string | null;
  durationSeconds?: number | null;
  caller?: User | null;
  callee?: User | null;
  createdAt: string;
}

export type InitiateCallBodyType =
  (typeof InitiateCallBodyType)[keyof typeof InitiateCallBodyType];

export const InitiateCallBodyType = {
  audio: "audio",
  video: "video",
  group: "group",
} as const;

export interface InitiateCallBody {
  calleeId?: number;
  chatId?: number;
  type: InitiateCallBodyType;
}

export type UpdateCallBodyStatus =
  (typeof UpdateCallBodyStatus)[keyof typeof UpdateCallBodyStatus];

export const UpdateCallBodyStatus = {
  active: "active",
  ended: "ended",
  declined: "declined",
} as const;

export interface UpdateCallBody {
  status: UpdateCallBodyStatus;
}

export interface SendGiftBody {
  receiverId: number;
  giftItemId: number;
  message?: string;
  isAnonymous?: boolean;
  chatId?: number;
}

export type StoryType = (typeof StoryType)[keyof typeof StoryType];

export const StoryType = {
  image: "image",
  video: "video",
  text: "text",
} as const;

export interface Story {
  id: number;
  userId: number;
  mediaUrl?: string;
  type: StoryType;
  text?: string | null;
  backgroundColor?: string | null;
  expiresAt: string;
  viewCount: number;
  isViewed: boolean;
  user?: User | null;
  createdAt: string;
}

export interface StoryGroup {
  user: User;
  stories: Story[];
  hasUnviewed: boolean;
}

export type CreateStoryBodyType =
  (typeof CreateStoryBodyType)[keyof typeof CreateStoryBodyType];

export const CreateStoryBodyType = {
  image: "image",
  video: "video",
  text: "text",
} as const;

export interface CreateStoryBody {
  mediaUrl?: string;
  type: CreateStoryBodyType;
  text?: string;
  backgroundColor?: string;
}

export interface UserStats {
  messagesSent: number;
  callsMade: number;
  callDurationSeconds: number;
  giftsSent: number;
  giftsReceived: number;
  chatsCount: number;
  contactsCount: number;
}

export type SearchUsersParams = {
  q: string;
};

export type GetMessagesParams = {
  chatId: number;
  limit?: number;
  before?: number;
};