# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username or name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a chat
      parameters:
        - name: chatId
          in: query
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    put:
      operationId: editMessage
      tags: [messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageBody"
      responses:
        "200":
          description: Edited message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Message deleted
  /messages/{messageId}/reactions:
    post:
      operationId: addReaction
      tags: [messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReactionBody"
      responses:
        "201":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
    delete:
      operationId: removeReaction
      tags: [messages]
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReactionBody"
      responses:
        "204":
          description: Reaction removed
...
  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get stories from contacts
      responses:
        "200":
          description: Stories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryGroup"
    post:
      operationId: createStory
      tags: [stories]
      summary: Create a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoryBody"
      responses:
        "201":
          description: Story created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{storyId}:
    delete:
      operationId: deleteStory
      tags: [stories]
      parameters:
        - name: storyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Story deleted
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        avatarColor:
          type: string
        status:
          type: string
          enum: [online, offline, away]
        lastSeen:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        isBot:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - avatarColor
        - status
        - isBot
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        status:
          type: string
          enum: [online, offline, away]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        text:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, video, audio, file, gift, call, sticker, system]
        mediaUrl:
          type: string
          nullable: true
        replyToId:
          type: integer
          nullable: true
        isEdited:
          type: boolean
        isRead:
          type: boolean
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        sender:
          $ref: "#/components/schemas/User"
          nullable: true
        replyTo:
          $ref: "#/components/schemas/Message"
          nullable: true
        giftData:
          $ref: "#/components/schemas/Gift"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - chatId
        - senderId
        - type
        - isEdited
        - isRead
        - reactions
        - createdAt
        - updatedAt
    SendMessageBody:
      type: object
      properties:
        chatId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [text, image, video, audio, file, sticker]
          default: text
        mediaUrl:
          type: string
        replyToId:
          type: integer
      required:
        - chatId
    EditMessageBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    Reaction:
      type: object
      properties:
        id:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        emoji:
          type: string
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - messageId
        - userId
        - emoji
        - createdAt
    AddReactionBody:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji
...
    GiftItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        animationType:
          type: string
          enum: [confetti, sparkle, hearts, fireworks, stars, balloons, roses, diamonds]
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
        stars:
          type: integer
        description:
          type: string
      required:
        - id
        - name
        - emoji
        - animationType
        - rarity
        - stars
        - description
    Gift:
      type: object
      properties:
        id:
          type: integer
        giftItemId:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        message:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        giftItem:
          $ref: "#/components/schemas/GiftItem"
          nullable: true
        sender:
          $ref: "#/components/schemas/User"
          nullable: true
        receiver:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - giftItemId
        - senderId
        - receiverId
        - isAnonymous
        - createdAt
...
    Story:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        mediaUrl:
          type: string
        type:
          type: string
          enum: [image, video, text]
        text:
          type: string
          nullable: true
        backgroundColor:
          type: string
          nullable: true
        expiresAt:
          type: string
        viewCount:
          type: integer
        isViewed:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - expiresAt
        - viewCount
        - isViewed
        - createdAt
    StoryGroup:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        hasUnviewed:
          type: boolean
      required:
        - user
        - stories
        - hasUnviewed
    CreateStoryBody:
      type: object
      properties:
        mediaUrl:
          type: string
        type:
          type: string
          enum: [image, video, text]
        text:
          type: string
        backgroundColor:
          type: string
      required:
        - type
...